Hema vardhan

+++ admin/src/components/ProductCard/ProductCard.jsx
@@ -3,8 +3,9 @@ import './ProductCard.css'
 import { useState } from 'react';
 import axios from 'axios';
 import { uploadFile } from '../../helper/uploadFile';
+// import ReactTimeAgo from 'react-time-ago'
 
-const ProductCard = ({ name, desc, disc, price, cat, dis_price, tags, image, avail, id, refresh }) => {
+const ProductCard = ({ name, desc, disc, price, cat, dis_price, tags, image, avail, id, refresh, isdeleted, date }) => {
     const [edit, setEdit] = useState(false);
     const [product, setProduct] = useState({
       name: name || '',
@@ -15,7 +16,8 @@ const ProductCard = ({ name, desc, disc, price, cat, dis_price, tags, image, ava
       category: cat || '',
       tags: tags || '',
       available: avail || false,
-      thumb_img: image 
+      thumb_img: image ,
+      isdeleted : isdeleted
     });
   
     const handleChange = async (e) => {
@@ -52,11 +54,23 @@ const ProductCard = ({ name, desc, disc, price, cat, dis_price, tags, image, ava
       }
     };
   
-    const handleDelete = async () => {
+    const handleDelete = async (e) => {
       try {
-        const resp = await axios.delete(`/product/${id}`);
-        if (resp.data.success) {
-          refresh();
+        const userConfirmed = window.confirm(`Are you sure you want to ${e.target.innerHTML} ${product.name}?`);
+        if(e.target.innerHTML === 'Recover'){
+          if (userConfirmed) {
+            const resp = await axios.get(`/product/${id}`)
+            if(resp.data.success){
+              refresh();
+            }
+          }
+        }else{
+          if (userConfirmed) {
+            const resp = await axios.delete(`/product/${id}`);
+            if (resp.data.success) {
+              refresh();
+            }
+          }
         }
       } catch (err) {
         console.log(err.message);
@@ -65,7 +79,7 @@ const ProductCard = ({ name, desc, disc, price, cat, dis_price, tags, image, ava
   
     return (
       <tr className='individual-product'>
-        <td>{edit ?<input type='file' name='thumb_img' onChange={handleChange} /> : <img src={product.thumb_img} alt='Product' className='product-image' />}</td>
+        <td className='product-image-holder'>{edit ?<input type='file' name='thumb_img' onChange={handleChange} /> : <img src={product.thumb_img} alt='Product' className='product-image' />}</td>
         <td>{edit ? <input type='text' value={product.name} name='name' onChange={handleChange} /> : product.name}</td>
         <td>{edit ? <input type='text' value={product.desc} name='desc' onChange={handleChange} /> : product.desc}</td>
         <td>{edit
@@ -82,7 +96,9 @@ const ProductCard = ({ name, desc, disc, price, cat, dis_price, tags, image, ava
         <td>{edit ? <input type='text' value={product.tags} name='tags' onChange={handleChange} /> : product.tags}</td>
         <td className='product-btn'>
           <button className='edit-btn' onClick={handleEdit}>{edit ? 'Submit' : 'Edit'}</button>
-          <button className='delete-btn' onClick={handleDelete}>Delete</button>
+          <button className='delete-btn' onClick={handleDelete}>{
+            product.isdeleted ? 'Recover' : 'Delete'  
+          }</button>
         </td>
       </tr>
     );
